+++ quantities/quantities.vcxproj
@@ -6,14 +6,14 @@
     <RootNamespace>quantities</RootNamespace>
     <ProjectName>quantities</ProjectName>
   </PropertyGroup>
-
   <Import Project="$(SolutionDir)principia.props" />
-
   <ItemGroup>
     <ClInclude Include="astronomy.hpp" />
     <ClInclude Include="bipm.hpp" />
     <ClInclude Include="cgs.hpp" />
     <ClInclude Include="constants.hpp" />
+    <ClInclude Include="dimensions.hpp" />
+    <ClInclude Include="dimensions_body.hpp" />
     <ClInclude Include="elementary_functions.hpp" />
     <ClInclude Include="elementary_functions_body.hpp" />
     <ClInclude Include="named_quantities.hpp" />
