+++ quantities/quantities.vcxproj.filters
@@ -65,6 +65,12 @@
     <ClInclude Include="parser_body.hpp">
       <Filter>Source Files</Filter>
     </ClInclude>
+    <ClInclude Include="dimensions.hpp">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="dimensions_body.hpp">
+      <Filter>Source Files</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="quantities_test.cpp">
